refactor to have his own controller
